#700 Fix current-navbar-item-highlight-not-getting-update-with-the-scroll-movement
