+++ packages/ui/src/components/header/MobileHeaderMenu.jsx
@@ -3,19 +3,18 @@ import { Link, useLocation, useNavigate } from "react-router-dom";
 import PropTypes from "prop-types";
 import { HEADER_ITEMS, LOGGEDIN_ITEMS } from "../../utils/Constants";
 import styles from "./MobileHeaderMenu.module.css";
-import { handleNavigation } from "../../utils/Helpers";
+import {
+  handleNavigation,
+  observeActiveSectionOnScroll,
+} from "../../utils/Helpers";
 import { AuthContext } from "../../contexts/Contexts";
+import { extractSectionIds } from "../../utils/Helpers";
 
 const MobileHeaderMenu = ({ closeMenu, isOpen }) => {
   const navigate = useNavigate();
   const { isAuthenticated } = useContext(AuthContext);
   const NAVBAR_ITEMS = isAuthenticated ? LOGGEDIN_ITEMS : HEADER_ITEMS;
-
-  const sectionIds = useMemo(() => {
-    return NAVBAR_ITEMS.filter((item) => item.type === "section").map(
-      (item) => item.url.split("#")[1] || ""
-    );
-  }, [NAVBAR_ITEMS]);
+  const sectionIds = useMemo(() => extractSectionIds(NAVBAR_ITEMS));
 
   const location = useLocation();
   const [activeSection, setActiveSection] = useState("");
@@ -35,33 +34,8 @@ const MobileHeaderMenu = ({ closeMenu, isOpen }) => {
   }, []);
 
   useEffect(() => {
-    const handleScroll = () => {
-      const scrollPosition = window.scrollY + window.innerHeight / 3;
-      let found = false;
-
-      for (let id of sectionIds) {
-        const section = document.getElementById(id);
-        if (section) {
-          const { offsetTop, offsetHeight } = section;
-          if (
-            scrollPosition >= offsetTop &&
-            scrollPosition < offsetTop + offsetHeight
-          ) {
-            setActiveSection(id);
-            found = true;
-            break;
-          }
-        }
-      }
-      if (!found && window.scrollY === 0) {
-        setActiveSection("");
-      }
-    };
-
-    window.addEventListener("scroll", handleScroll);
-    handleScroll();
-
-    return () => window.removeEventListener("scroll", handleScroll);
+    const cleanup = observeActiveSectionOnScroll(sectionIds, setActiveSection);
+    return cleanup;
   }, [sectionIds]);
 
   if (!isOpen) return null;
